+++ packages/terra-dropdown-button/CHANGELOG.md
@@ -4,6 +4,11 @@
 * Fixed
  * Fixed dropdown button list items of not announcing the selection state.
 
+## 1.33.0 - (April 27, 2023)
+
+* Fixed
+   * Fixed selection logic to cycle from last option to first option and vice-versa
+
 ## 1.32.0 - (March 1, 2023)
 
 * Changed